clean imports and fixed security

+++ app/services/comment.py
@@ -1,10 +1,6 @@
 from uuid import UUID
-from typing import Annotated, List
-from fastapi.routing import APIRouter
-from fastapi import HTTPException, status, Depends
-from sqlalchemy import select
+from fastapi import HTTPException, status
 from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy.orm import selectinload
 
 from app.models.comment import Comment
 from app.models.user import User
@@ -16,7 +12,9 @@ from app.repositories.comment import (
     get_my_comments_db,
     update_comment_db,
 )
+from app.repositories.notification import create_notification_db
 from app.schemas.comment import CommentCreate, CommentUpdate
+from app.services.notification import create_notification_service
 from app.services.post import get_post_service
 
 
@@ -41,6 +39,10 @@ async def create_comment_service(
     new_comment = await create_comment_db(comment, db)
     loaded_comment = await get_comment_by_id_db(new_comment.id, db)
 
+    if current_user.id != post.author.id:
+        notification = create_notification_service(current_user, post, loaded_comment)
+        await create_notification_db(notification, db)
+
     return loaded_comment
 
 
@@ -57,17 +59,31 @@ async def get_comment_service(comment_id: UUID, db: AsyncSession):
 
 
 async def update_comment_service(
-    form_data: CommentUpdate, comment_id: UUID, db: AsyncSession
+    form_data: CommentUpdate, comment_id: UUID, db: AsyncSession, current_user: User
 ):
     comment = await get_comment_by_id_db(comment_id, db)
+
+    if comment.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You cannot update this comment",
+        )
+
     data = form_data.model_dump(exclude_unset=True)
     updated_comment = await update_comment_db(data, comment, db)
 
     return updated_comment
 
 
-async def delete_comment_service(comment_id: UUID, db: AsyncSession):
+async def delete_comment_service(
+    current_user: User, comment_id: UUID, db: AsyncSession
+):
     comment = await get_comment_by_id_db(comment_id, db)
+    if current_user.id != comment.user_id and current_user.id != comment.post.author.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You are not allowed to delete this comment.",
+        )
 
     await db.delete(comment)
     await db.commit()
